+++ app/app.module.ts
@@ -2,13 +2,14 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
+import { HttpModule } from '@angular/http';
 
 //my-app local
 import { AppComponent } from './app.component';
 import { UserComponent } from './components/user.component';
 
 @NgModule({
-  imports: [BrowserModule, FormsModule],
+  imports: [BrowserModule, FormsModule, HttpModule],
   declarations: [AppComponent, UserComponent],
   bootstrap: [AppComponent]
 })